Asset extraction:
- Add PM64AudioFactory: byte-swap SBN container and all sub-formats (BGM, SEF, BK, PER, PRG, MSEQ, INIT) from big-endian to native-endian at extraction time
- Add PM64MessageFactory: extract message data blob (runtime byte-swap of u32 offsets in dma_load_msg)
- Add PM64StoryImageFactory: extract CI8/IA8 intro story page images with palettes
- Add charset asset definitions (fonts, postcard, letter content images and palettes)
- Add sprite asset definitions for all NPC and player sprites
- Add theater asset definitions (walls, curtains, floor textures and vertices)
- Upgrade audio.yml to use PM64:AUDIO factory instead of raw BLOB

Audio system:
- Implement audio thread with mutex/condvar synchronization, generating 2 audio frames per game frame (60Hz)
- Add software RSP audio mixer (SSE2/NEON) ported from libstarship
- Remove BE16SWAP/BE32SWAP calls from au_load_INIT, au_load_song_files, au_reload_song_files, au_ambient_load since data is now native-endian
- Add runtime byte-swap of BGM SeqArgs (multi-byte command fields read from big-endian stream)
- Initialize BGM tracks with defaultInstrument instead of NO_INSTRUMENT to avoid crash when audio thread races game thread
- Add InstrumentBinary struct (0x30 bytes) for unpacking N64 BK instrument records into native Instrument structs on 64-bit

64-bit compatibility fixes:
- Widen Evt struct fields (functionTemp, varTable, loopStartTable, switchBlockValue) from s32 to intptr_t/Bytecode for 64-bit pointer storage
- Fix evt_get_variable/evt_set_variable return types to Bytecode
- Fix msg_get_printer_for_msg/msg_printer_load_msg msgID parameter from s32 to intptr_t
- Add MSG_ID_IS_PTR macro replacing sign-check for distinguishing message IDs from buffer pointers
- Fix AU_FILE_RELATIVE macro to use intptr_t instead of s32 for pointer arithmetic
- Fix ALDMAproc2 typedef to use intptr_t for addr parameter
- Fix sprite code to use SpriteAnimData struct fields instead of u32* pointer arithmetic (pointers are 8 bytes on 64-bit)
- Disable spr_load_npc_extra_anims which corrupts heap with 64-bit pointer assumptions
- Fix bgm_player.c pointer casts from (s32) to (intptr_t) for phraseStartPos

Graphics and rendering:
- Add unified frame orchestration (gfx_frame.c) with double-buffered GfxPool, eliminating flickering from multiple buffer swaps
- Consolidate background and main display list building into single submission via Graphics_ThreadUpdate
- Add GBI middleware for OTR path resolution in static display lists
- Register DisplayList and Vertex resource factories for shape model loading
- Refactor shape_loader.c to load display lists from OTR resources with G_VTX_OTR_HASH vertex resolution
- Extend PM64ShapeFactory to parse and export display lists with N64-to-file-offset vertex conversion
- Add display list context tracking for debugging (GameEngine_SetDisplayListContext)

Game code port changes:
- Replace DMA_COPY_SEGMENT calls in load_engine_data with no-ops (data loaded from assets)
- Replace dma_copy/load_font_data in msg.c with memcpy from LOAD_ASSET
- Replace dma_copy in intro story images (hos_05) with ResourceGetDataByName
- Convert curtains.c from inline INCLUDE_IMG/vtx data to OTR asset loading
- Add narrator.c files for hos_04, hos_05, hos_10 story sequences
- Add init_globals.c for BSS zero-initialization (prevents crashes from garbage values like curPartner=70)
- Add gfx_pool.h with GfxPool struct and master display list buffer

+++ src/game/audio/core.h
@@ -7,7 +7,7 @@
 void create_audio_system(void);
 //void nuAuPreNMIFuncSet(NUAuPreNMIFunc func);
 void nuAuMgr(void* arg);
-s32 nuAuDmaCallBack(s32 addr, s32 len, void *state, u8 useDma);
+intptr_t nuAuDmaCallBack(intptr_t addr, s32 len, void *state, u8 useDma);
 //ALDMAproc nuAuDmaNew(NUDMAState** state);
 //void nuAuCleanDMABuffers(void);
 //void nuAuPreNMIProc(NUScMsg mesg_type, u32 frameCounter);
@@ -69,7 +69,7 @@ void au_load_PER(AuGlobals* globals, s32 romAddr);
 void au_load_PRG(AuGlobals* arg0, s32 romAddr);
 InstrumentBank* au_get_BK_instruments(BankSet bankSet, u32 bankIndex);
 BKFileBuffer* au_load_BK_to_bank(s32 bkFileOffset, BKFileBuffer* bkFile, s32 bankIndex, BankSet bankSet);
-void au_swizzle_BK_instruments(s32 bkFileOffset, BKFileBuffer* bkFile, InstrumentBank instruments, u32 instrumentCount, u8 arg4);
+void au_swizzle_BK_instruments(intptr_t bkFileOffset, BKFileBuffer* bkFile, InstrumentBank instruments, u32 instrumentCount, u8 arg4);
 BKFileBuffer* au_load_static_BK_to_bank(s32* bkFileOffset, void* vaddr, s32 bankIndex, BankSet bankSet);
 s32 au_load_aux_bank(s32 bkFileOffset, s32 bankIndex);
 void au_clear_instrument_group(s32 bankIndex, BankSet bankSet);
